refactor mode_info
