+++ app/api/model_info.py
@@ -4,15 +4,9 @@ from slowapi.util import get_remote_address
 
 from ..services.prediction import PneumoniaPredictionService
 from ..core.logger import get_logger
-from ..docs.sections.prediction_metadata import PredictionMetadata
 from ..utils.get_prediction_service import get_prediction_service
 from ..docs.sections.model_info_metadata import ModelInfoMetadata
-from ..models.schemas import (
-    ModelInfoResponse,
-    ModelInfoValidationErrorResponse,
-    ModelInfoNotFoundResponse,
-    ModelInfoServiceUnavailableResponse
-)
+
 
 logger = get_logger(__name__)
 router = APIRouter()
@@ -23,16 +17,8 @@ model_info_metadata = ModelInfoMetadata.get_metadata()
 
 @router.get(
     "/model/info",
-    response_model=ModelInfoResponse,
     tags=["Model"],
-    responses={
-        404: {"model": ModelInfoNotFoundResponse},
-        422: {"model": ModelInfoValidationErrorResponse}, 
-        503: {"model": ModelInfoServiceUnavailableResponse}
-    },
-    summary=model_info_metadata.get("summary", ""),
-    description=model_info_metadata.get("description", ""),
-    response_description=model_info_metadata.get("response_description", "")
+    **model_info_metadata,
 )
 async def get_model_info(
     prediction_service: PneumoniaPredictionService = Depends(get_prediction_service)
