+++ app/docs/sections/model_info_metadata.py
@@ -1,4 +1,12 @@
 from typing import Dict, Any
+from ...docs.base_builder import build_response
+from ...models.error_codes import ErrorCode
+from ...models.model_info_schemas import (
+    ModelInfoResponse,
+    ModelInfoValidationErrorResponse,
+    ModelInfoNotFoundResponse,
+    ModelInfoServiceUnavailableResponse
+)
 
 
 class ModelInfoMetadata:
@@ -70,98 +78,74 @@ architecture details, performance metrics, and training statistics.</p>
 
     @staticmethod
     def get_200_response() -> Dict[str, Any]:
-        return {
-            "description": "Model information retrieved successfully",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "$ref": "#/components/schemas/ModelInfoResponse"
-                    },
-                    "examples": {
-                        "model_loaded": {
-                            "summary": "Model successfully loaded",
-                            "value": {
-                                "loaded": True,
-                                "model_path": "models/pneumonia_model_efficientnet_b0.onnx",
-                                "input_name": "input",
-                                "output_name": "output",
-                                "mean": 0.480,
-                                "std": 0.237,
-                                "target_size": [192, 192],
-                                "labels": ["NORMAL", "PNEUMONIA"],
-                                "model_type": "efficientnet_b0"
-                            }
-                        },
-                        "model_not_loaded": {
-                            "summary": "Model not loaded",
-                            "value": {
-                                "loaded": False
-                            }
-                        }
+        return build_response(
+            description= "Model information retrieved successfully",
+            model= ModelInfoResponse,
+            examples= {
+                "model_loaded": {
+                    "summary": "Model successfully loaded",
+                    "value": {
+                        "loaded": True,
+                        "model_path": "models/pneumonia_model_efficientnet_b0.onnx",
+                        "input_name": "input",
+                        "output_name": "output",
+                        "mean": 0.480,
+                        "std": 0.237,
+                        "target_size": [192, 192],
+                        "labels": ["NORMAL", "PNEUMONIA"],
+                        "model_type": "efficientnet_b0"
+                    }
+                },
+                "model_not_loaded": {
+                    "summary": "Model not loaded",
+                    "value": {
+                        "loaded": False
                     }
                 }
             }
-        }
+        )
         
     @staticmethod
     def get_404_response() -> Dict[str, Any]:
-        return {
-            "description": "Model not found",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "$ref": "#/components/schemas/ModelInfoNotFoundResponse"
-                    },
-                    "example": {
-                        "detail": "Model 'invalid_model' not found",
-                        "error_code": "MODEL_NOT_FOUND",
-                        "available_models": ["standard", "efficientnet_b0"],
-                        "timestamp": "2025-09-13T10:30:00.000Z"
-                    }
-                }
+        return build_response(
+            description= "Model not found",
+            model= ModelInfoNotFoundResponse,
+            example= {
+                "detail": "Model 'invalid_model' not found",
+                "error_code": ErrorCode.MODEL_NOT_FOUND,
+                "available_models": ["standard", "efficientnet_b0"],
+                "timestamp": "2025-09-13T10:30:00.000Z"
             }
-        }
+        )
         
     @staticmethod
     def get_503_response() -> Dict[str, Any]:
-        return {
-            "description": "Prediction service unavailable",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "$ref": "#/components/schemas/ModelInfoServiceUnavailableResponse"
-                    },
-                    "example": {
-                        "detail": "Prediction service not available",
-                        "error_code": "SERVICE_UNAVAILABLE",
-                        "service_status": "not_initialized",
-                        "timestamp": "2025-09-13T10:30:00.000Z"
-                    }
-                }
+        return build_response(
+            description= "Prediction service unavailable",
+            model= ModelInfoServiceUnavailableResponse,
+            example= {
+                "detail": "Prediction service not available",
+                "error_code": ErrorCode.SERVICE_UNAVAILABLE,
+                "service_status": "not_initialized",
+                "timestamp": "2025-09-13T10:30:00.000Z"
             }
-        }
+        )
         
     @staticmethod
     def get_422_response() -> Dict[str, Any]:
-        return {
-            "description": "Validation Error",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "$ref": "#/components/schemas/ModelInfoValidationErrorResponse"
-                    },
-                    "example": {
-                        "detail": [
-                            {
-                                "loc": ["query", "model"],
-                                "msg": "Invalid model type specified",
-                                "type": "value_error"
-                            }
-                        ]
+        return build_response(
+            description= "Validation Error",
+            model= ModelInfoValidationErrorResponse,
+            example= {
+                "detail": [
+                    {
+                        "loc": ["query", "model"],
+                        "msg": "Invalid model type specified",
+                        "type": "value_error"
                     }
-                }
+                ]
             }
-        }
+        )
         
     @classmethod
     def get_full_description(cls) -> str:
