+++ app/models/error_codes.py
@@ -12,3 +12,4 @@ class ErrorCode(str, Enum):
     MODEL_NOT_LOADED = "MODEL_NOT_LOADED"
     PREDICTION_FAILED = "PREDICTION_FAILED"
     INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR"
+    MODEL_NOT_FOUND = "MODEL_NOT_FOUND"
